fix: ✏️ 사용자 로그 관리 정책 변경에 따른 Device API 수정 (#104)
* fix: 배포 파이프라인 이미지 빌드 버전 추가

* fix: device entity os & model 필드 제거

* fix: device token dto model, os 필드 제거

* fix: device_dto to_entity() 메서드 복구

* fix: device update 메서드 호출 시, activate 상태로 변경

* fix: device 등록 비지니스 로직 수정

* test: device controller unit test에서 model, os 제거

* fix: device_fixture model, os 필드 제거

* rename: fixture token 변경 상수명 수정

* fix: device entity 조회 시 활성화 조건 추가

* fix: device usecase 테스트 시나리오 수정

* fix: device 쿼리 자동 조건문 수정

* test: origin_token에 대한 정보가 없는 경우 new_token 등록

* fix: @sql_restriction 제거

* test: 수정 요청 & 기존 토큰 없을 시, 수정 토큰으로 신규 등록 테스트

* fix: device_register_service 로직 수정

* fix: device_dto new_token 필드 제거

* test: device controller unit test 수정

* fix: device 응답 객체 생성 시 요청 토큰이 아닌 서버 토큰을 응답으로 사용

* fix: device_register_service 제거 && 로직 단순화

* fix: 사용자가 보낸 토큰이 서버에서 비활성화 상태일 때 예외처리

* test: test case 요청 수정

* rename: device -> device_token으로 수정

* rename: device_token entity to_string 수정

* rename: device token put 요청 응답 key 수정

* docs: device put api 스웨거 수정

* fix: device controller 경로 수정

* test: device_token controller unit test 변경된 url 및 응답 포맷 반영

* test: device_fixture -> device_token_fixture

* test: 단일 token에 대한 시나리오로 테스트 수정 및 비활성화 토큰 요청 에러 케이스 추가

* fix: register_device_token 파사드 패턴 적용

* fix: device_token_unregister 파사드 패턴 적용

* test: device token 단위 테스트 분리

* test: device_token_unregister_service_test 분리

* fix: user_account_use_case 파사드 패턴 적용

* test: user_account_usecase unit test 분리

* fix: device entity os & model 필드 제거

* fix: device token dto model, os 필드 제거

* fix: device_dto to_entity() 메서드 복구

* fix: device update 메서드 호출 시, activate 상태로 변경

* fix: device 등록 비지니스 로직 수정

* test: device controller unit test에서 model, os 제거

* fix: device_fixture model, os 필드 제거

* rename: fixture token 변경 상수명 수정

* fix: device entity 조회 시 활성화 조건 추가

* fix: device usecase 테스트 시나리오 수정

* fix: device 쿼리 자동 조건문 수정

* test: origin_token에 대한 정보가 없는 경우 new_token 등록

* fix: @sql_restriction 제거

* test: 수정 요청 & 기존 토큰 없을 시, 수정 토큰으로 신규 등록 테스트

* fix: device_register_service 로직 수정

* fix: device_dto new_token 필드 제거

* test: device controller unit test 수정

* fix: device 응답 객체 생성 시 요청 토큰이 아닌 서버 토큰을 응답으로 사용

* fix: device_register_service 제거 && 로직 단순화

* fix: 사용자가 보낸 토큰이 서버에서 비활성화 상태일 때 예외처리

* test: test case 요청 수정

* rename: device -> device_token으로 수정

* rename: device_token entity to_string 수정

* rename: device token put 요청 응답 key 수정

* docs: device put api 스웨거 수정

* fix: device controller 경로 수정

* test: device_token controller unit test 변경된 url 및 응답 포맷 반영

* test: device_fixture -> device_token_fixture

* test: 단일 token에 대한 시나리오로 테스트 수정 및 비활성화 토큰 요청 에러 케이스 추가

* fix: register_device_token 파사드 패턴 적용

* fix: device_token_unregister 파사드 패턴 적용

* test: device token 단위 테스트 분리

* test: device_token_unregister_service_test 분리

* fix: user_account_use_case 파사드 패턴 적용

* test: user_account_usecase unit test 분리

* fix: user_profile_update_dto merge conflict

+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/users/service/UserDeleteService.java
@@ -1,6 +1,8 @@
 package kr.co.pennyway.api.apis.users.service;
 
 import kr.co.pennyway.domain.domains.oauth.service.OauthService;
+import kr.co.pennyway.domain.domains.user.exception.UserErrorCode;
+import kr.co.pennyway.domain.domains.user.exception.UserErrorException;
 import kr.co.pennyway.domain.domains.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +24,11 @@ public class UserDeleteService {
     private final OauthService oauthService;
 
     @Transactional
-    public void deleteUser(Long userId) {
+    public void execute(Long userId) {
+        if (!userService.isExistUser(userId)) throw new UserErrorException(UserErrorCode.NOT_FOUND);
+
+        // TODO: [2024-05-03] 하나라도 채팅방의 방장으로 참여하는 경우 삭제 불가능 처리
+        
         oauthService.deleteOauthsByUserId(userId);
         userService.deleteUser(userId);
     }
